Move ConfirmedCasesDetailsTable to components/index/cards/charts
