+++ components/index/cards/charts/ConfirmedCasesDetailsTable.vue
